rafactor: create useMoviePlayer hook
